+++ apps/riscv-tui/src/ui/component/mid/mod.rs
@@ -3,6 +3,6 @@ pub mod csr;
 pub mod register;
 
 #[cfg(not(feature = "zicsr"))]
-const MID_TITLE: &str = "Registers";
+const MID_TITLE: &str = "Reg (H) Dec/Hex";
 #[cfg(feature = "zicsr")]
-const MID_TITLE: &str = "Reg / Csr (Press C)";
+const MID_TITLE: &str = "(C) Reg / Csr (H) Dec/Hex ";